fix wrong files
